День 5: ...

+++ Product.cs
@@ -29,7 +29,7 @@ namespace StroyMat
         public string Description { get; set; }
         public string Image { get; set; }
         public Nullable<int> ProductionPersonCount { get; set; }
-        public Nullable<int> ProductionWorkshopNumber { get; set; }
+        public Nullable<double> ProductionWorkshopNumber { get; set; }
         public decimal MinCostForAgent { get; set; }
     
         public virtual ProductType ProductType { get; set; }
